RTSS-216 update bulk unit import for the numerator/denominator conversion split
`unit_value.conversionFactor`/`unit_conversion.factor` were replaced by a whole-number `unitsOfBasePerUnit: Long` on unit values and a `factorNumerator`/`factorDenominator` ratio pair on unit conversions (see `unitvalue/README.md` and `unitconversion/README.md` for why unit values collapse to one integer while conversions need a true ratio). Bulk import still spoke the old single-`BigDecimal`-factor contract: `UnitConversionService.bulkInsertValidatedList` and `UnitValueValidator.validateUnitValueUpdate` referenced the removed `factor`/`conversionFactor` fields, and `UnitValueBulkInsertDto`, `UnitConversionBulkInsertDto`, and `UnitValueBulkSaveEntry` were typed against the old model.

`UnitValueBulkInsertDto.unitsOfBasePerUnit` is now `Long`; `UnitConversionBulkInsertDto` has `numerator`/`denominator` longs instead of a single factor; `UnitValueBulkSaveEntry` matches `UnitValueEntity`'s field. `BulkUnitImportValidator`'s unit-value rule requires `unitsOfBasePerUnit >= 1` (a positive whole number, not just non-zero) to match the new invariant, and its conversion rule validates `numerator`/`denominator` independently rather than one derived factor. `BulkUnitImportService` passes the renamed/retyped fields through with no change to save ordering, Piece auto-provisioning, or cache batching. `UnitValueValidator`'s blank-checks on `name`/`code` now use `StringUtils.hasValue` instead of `.isNullOrBlank()`, consistent with the rest of this domain.

`BulkUnitImportServiceTest` gained coverage for the validation-failure path (throws with the full error list, makes no calls to any of the group/value/conversion services) and for `saveConversions` (resolves `fromUnitCode`/`toUnitCode` to ids and passes `numerator`/`denominator` through unchanged). The `/commit-message` command now says explicitly that generated messages describe the code and its history, not the agent session that produced it.

+++ src/main/kotlin/me/ezra_home/retail_software_solution/organizations/business/unit_bulk_import/api/UnitValueBulkInsertDto.kt
@@ -5,5 +5,5 @@ data class UnitValueBulkInsertDto(
     val code: String?,
     val description: String?,
     val baseUnitCode: String?,
-    val unitsOfBasePerUnit: Double?
+    val unitsOfBasePerUnit: Long?
 )
